+++ OMG.PSUtilities.Core/OMG.PSUtilities.Core.psm1
@@ -20,7 +20,7 @@ Get-ChildItem -Path "$PSScriptRoot\Public\*.ps1" -Recurse | ForEach-Object {
 $PublicFunctions = @(
     'Find-PSUFilesContainingText'
     'Get-PSUConnectedWifiInfo'
-    'Get-PSUFunctionHelpInfo'
+    'Get-PSUFunctionCommentBasedHelp'
     'Get-PSUGitRepositoryChanges'
     'Get-PSUInstalledSoftware'
     'Get-PSUUserSession'