+++ OMG.PSUtilities.Core/CHANGELOG.md
@@ -3,3 +3,8 @@
 ## [1.0.0] - 2025-07-16
 - Initial scaffolding for OMG.PSUtilities.Core
 
+
+## [1.0.3] - 2025-07-17
+- [Deleted File/Folder] [Get-PSUFunctionHelpInfo.ps1] : Renamed to Get-PSUFunctionCommentBasedHelp.ps1
+- [Modified File] [Get-PSUGitRepositoryChanges.ps1] : Fixed a minor bug
+- [New File] [Get-PSUFunctionCommentBasedHelp.ps1] : This function pulls out a specific help section (like .SYNOPSIS or .EXAMPLE) from a PowerShell script, useful for reading documentation automatically.